refactor(transcription): move manager to application layer, split into SRP submodules
Pure service (no Dioxus), was a ui-layer violation. Split the 502-line module into job/processing/append. Public API re-exported, importers untouched.
